Move moments/test_moments_rolling_*.py tests to windows/test_rolling_*.py
