+++ pandas/tests/window/test_rolling_functions.py
